(8b50589691e64845927bbe22c89674c56303bae1) 0.9beta10 - fix crash when unpausing after starting squeezelite before strm_s received (573061c220de5821b7957473dd89b88b21d3be69) 0.9beta9 - support for 384000, 352800 sample rates - ensure windows version recovers from disconnect - move some warn to info to hide debugging on stream disconnect - fix costmetic issue displaying latency requested value for portaudio Increased squeezelite revision to 201.

+++ main.c
@@ -149,7 +149,7 @@ int main(int argc, char **argv) {
 				if (m) alsa_mmap = atoi(m);
 #endif
 #if PORTAUDIO
-				pa_latency = atoi(optarg);
+				pa_latency = (unsigned)atoi(optarg);
 #endif
 			}
 			break;
